requirements.txt
pandas==2.2.2
selenium==4.22.0
beautifulsoup4==4.12.3
scrapegraphai==1.9.0
python-dotenv==1.0.1